Application: Added helper method to calculate score and removed helper methods from application helper
