+++ app/helpers/application_helper.rb
@@ -1,11 +1,4 @@
 # frozen_string_literal: true
 
 module ApplicationHelper
-  def attempted_count
-    session[:question_answer_ids] ? session[:question_answer_ids].length : 0
-  end
-
-  def questions_count
-    Question.count
-  end
 end
